Update init script to account for new statsd tcp config.
When mounting an empty volume into the container at /opt/statsd, the init
scripts only copy config_udp.js into /opt/statsd/config.js. This causes
an error with stats, which is expecting configs with protocol in the name,
and it prevents the new config_tcp.js from being installed.