Fix incorrect port in GRAPHITE_URL configuration.
The GRAPHITE_URL configuration in carbon.conf uses port 80, but the
graphite webapp provided in the container listens on port 8080. When
this setting is incorrect, carbon is unable to perform tagging of
incoming metrics with an error similar to the following:

  [console] Error tagging my.metric.name: Connection was refused by other side: 111: Connection refused.

This patch updates the configuration to use the proper port.